refactor(intent): move tests to internal/intent package (GH-803)
- Create internal/intent package with Intent type and detection logic
- Export helper functions (IsGreeting, IsQuestion, IsTask, etc.) for testing
- Update telegram package to re-export from internal/intent for backward compat
- Move all test functions from telegram/intent_test.go to intent/intent_test.go
